update list address observe updates

+++ app/src/main/java/com/example/alfaresto_customersapp/ui/components/trackOrder/TrackOrderViewModel.kt
@@ -179,7 +179,7 @@ class TrackOrderViewModel @Inject constructor(
     }
 
     fun getTimeEstimation(duration: Double): String {
-        return (duration / 60.00).toString() + " min"
+        return (duration / 60.00).toInt().toString() + " min"
     }
 
 }
